add JS target
